pyproject.toml

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "urban_renovation"
version = "0.1.0"
description = "Urban Renovation Literature Extraction Tool"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "python-docx>=1.1.0",
    "python-dotenv>=1.0.0",
    "Jinja2>=3.1.0",
]
requires-python = ">=3.13,<3.14"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "pillow>=10.0.0",
    "pymupdf>=1.24.0",
    "reportlab>=4.0.0",
]
report = [
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "pillow>=10.0.0",
    "pymupdf>=1.24.0",
    "reportlab>=4.0.0",
]
topic-discovery = [
    "bertopic>=0.16.0",
    "sentence-transformers>=3.0.0",
    "umap-learn>=0.5.7",
    "hdbscan>=0.8.40",
    "scikit-learn>=1.5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
